refactor(tests): merge duplicate UIEffects suites into ui-effects.host.test.ts
